Fix bug in Reply.java

+++ portfolio/src/main/java/com/google/sps/commentart/Reply.java
@@ -3,7 +3,7 @@ package com.google.sps.commentart;
 public class Reply {
   private String responderName;
   private String response;
-  private int sentiment;
+  private float sentiment;
 
   public Reply(String responderName, String response, float sentiment) {
     this.responderName = responderName;
